% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezbayes-package.R
\docType{package}
\name{ezbayes-package}
\alias{ezbayes}
\alias{ezbayes-package}
\title{ezbayes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnsonpc2/ezbayes}
  \item Report bugs at \url{https://github.com/johnsonpc2/ezbayes/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
